update to FG_GIRLS do files
Two files that code FGC indicators among girls are available, one that uses only the BR file (FG_GIRLS.do) and one that requires both the IR and BR file and contains a merge (FG_GIRLS_merge.do).

If all the information about the daughters female circumcision is found in the BR file, then the merge with the IR is not needed and the user should use the FG_GIRLS.do file. Otherwise use FG_GIRLS_merge.do file.
Please see the notes in the main file and these do files.